+++ ext/ClimaCouplerCMIPExt.jl
@@ -16,7 +16,13 @@ module ClimaCouplerCMIPExt
 
 import ClimaCoupler
 import ClimaCoupler:
-    Checkpointer, FieldExchanger, FluxCalculator, Interfacer, Utilities, Plotting
+    Checkpointer,
+    FieldExchanger,
+    FluxCalculator,
+    Interfacer,
+    TimeManager,
+    Utilities,
+    Plotting
 import Oceananigans as OC
 import ClimaOcean as CO
 import ClimaSeaIce as CSI
@@ -24,14 +30,15 @@ import ClimaCore as CC
 import ClimaParams as CP
 using KernelAbstractions: @kernel, @index, @inbounds
 
-# Include helper functions first (used by both oceananigans.jl and clima_seaice.jl)
-include("ClimaCouplerCMIPExt/climaocean_helpers.jl")
-
 # Include skin temperature utilities
 include("ClimaCouplerCMIPExt/skin_temperature.jl")
 
-# Include the model files
+# Include the model files first so their types are available to climaocean_helpers.jl
 include("ClimaCouplerCMIPExt/oceananigans.jl")
 include("ClimaCouplerCMIPExt/clima_seaice.jl")
+include("ClimaCouplerCMIPExt/climaocean_helpers.jl")
+
+include("ClimaCouplerCMIPExt/ocean_diagnostics.jl")
+include("ClimaCouplerCMIPExt/seaice_diagnostics.jl")
 
 end # module ClimaCouplerCMIPExt